db/seed/requirements.txt
# ==============================================================================
# 시드 스크립트 의존성
# ------------------------------------------------------------------------------
# 대부분 메인 프로젝트 requirements.txt에 이미 포함되어 있을 가능성이 높습니다.
# 누락된 경우에만 설치하세요:
#     pip install -r db/seed/requirements.txt
#
# 또는 메인 requirements.txt에 이 항목들을 병합하는 것을 권장합니다.
# ==============================================================================

asyncpg>=0.29.0
redis>=5.0.0
chromadb>=0.5.0
python-dotenv>=1.0.0
